Split the examples into 2 different ones
